Dev: clj-kondo: export config
It is a bit weird tbh.
Not sure this is the recommended way to do it but it seems to work.

The exported config is the base source of truth, and it's pulled into this
project with
1. `clj-kondo --copy-configs --dependencies --lint "$(clojure -Spath)"`
2. adding `:config-paths`

Users of this library can do the same for their projects.

PS. other options (more generic) options for solving this would be:
* symlinking
* copying