Significant refactor of jlink.clj
Rewrote the internals to make the code simpler and, hopefully, clearer.
The code could still be simpler but this was what I had time for. I took
advantage of the 'easier' ways to specify defaults such that the bridge
works on my machines without environment variables.

I also started a code glossary and notes files for future use.