Fix #29: - should become Minus for 1 arg, Subtract for more
Problem: Clojure devs are used to `-` taking any number of args. But we map that to `Subtact`, which only takes two.

Fix: Decide whether to use Minus or Subtract based on the number of arguments, and throw if too many. Do this during convert, by adding an _experimental_ support for having a fn as the alias value. Also modify wolfram ns to keep the original symbols (e.g. - for -) so the conversion still is done by convert, plus add docstrings there for these.

Cleanup: rm `(def fn ..)` from the wolfram ns, as we are adding a macro of the same name there.