[Enhancement] Include custom aliases in generated wolfram ns
Issue: `wl/init!` / `wl/eval` make it possible to teach Wolfram about custom aliases, but they only work in the symbolic form, i.e. `(list '<the symbol> args...)`. It would be more convenient to have function vars for this as we do for the built-in aliases such as `*`, so that the user could do `(w/<the symbol> args...)`.

Fix: Extend write-ns! to take a map of opts, including aliases, and write these to the file.

Fixes #59 partly.