Fix #68: Clean up, document core ns; correct license to MLP
Code/doc cleanup: Hide vars that users should not care about, make sure the rest is well documented.

License: Reading the core ns I've noticed the original license was Mozilla P. L. v1.1. Reportedly, it is ok to upgrade to a newer version (here 2`.0) and the key difference between Eclipse and Mozilla P.L. is that the latter is compatible with GPL (and, in v2, with Apache license).

For diff between MLP 1.1 and 2.0, see https://www.mozilla.org/en-US/MPL/2.0/Revision-FAQ/#what-has-changed