Set up docs generation with Clay (#67)
ssue: We have one clay/kindly notebook. We want to adopt Clay (and later also Quarto) to render all of our docs, to be compatible with other scicloj libs.

Fix: Update clay config, add notebooks/ as the root of the new docs. Add a clojure.tools build script to build the whole site.

* Move the docs from generated/ directly to docs/ for compatibility w/ GH Pages
* :dev deps: updated clay, removed obsolete orchard

---------

Co-authored-by: daslu <daniel.slutsky@gmail.com>