Fix #79 Add core/restart
Issue: We want to be able to restart the kernel with new aliases, without restarting the REPL.

Fix: Make stop not only terminate the kernel, but also clean up everything so that start will start from scratch. Also, make it not fail if the kernel is not actually started yet.

Also make terminate do call .close as docs suggest we should, and make start retry for a few times with a growing delay - since close+terminate make take a while to finish and
may throw before that.