Fix core-test for packages - we need eval; fix restart
Running this code would work in the REPL, where it is built sexp by sexp, but does not work in a non-interactive contex, where we try to read the whole file first, and fail b/c the dynamically created namespaces are not created yet (as that only happens during eval) => combine quote + eval to postpone until eval time.

For restart test, we clash now with an alias already included in the defaults, so we needed to switch to a different one.