Fix: Support quotes, evaluated code inside w/fn
Issue: We used to take the whole body of w/fn as unevaluated, i.e. effectively quoted sexp. This was to avoid linters complainng about unknown symbols for the arguments. However, we actually want to be able to use evaluated code (e.g. quick in-line, in-clj computation) and only use quotes where necessary.

Fix: Include the body as-is without quoting it, but do walk it and turn any references to the argument symbols into quoted symbols.

Also: Add tests, improve code structure, be explicit about code branches that do not work, may convert more eager to fail early. `->wl` - add missing passing on of jlink-instance.

Fixes #106