Refactor how we force string interpretation as Wolfram expr
Issue: It is not obvious when to interpret string as a primitive Wolfram string and when as a Wolfram expression. We used to force the latter by turning it into `(quote <str>)` and having special handling of that - but it was not obvious.

Fix: Be explicit - 1) added `convert/->wolfram-str-expr` to centralize the marking of a string as a Wolfram expr.; 2) Replaced `(quote str)` with a map with a well named key, namely `{::wolfram-str-expr str}`.