Graphics resizing: pull out fns, add Show on first call too
Ensure the graphics is scaled to the window already on the first render.

The math canvas provides no size info, so we need to set a min. size, which works on the JFrame (but not, it seems, if done to the `math`). Then we need to call `frame.pack` _before_ we run any code relying on any of these sizes - in particular the code setting the resized wl-expr.