Graphics code changes (#148)
# 1. Make Swing the new default Java graphics impl.
    
Move former `graphics` to wolframite.legacy.tools.graphics-awt and turn `wolframite.tools.experimental` to `...graphics`.
    
The advantage is that it uses a more modern (though also ancient) graphics library, and that more work is done by Wolfram provided code. It also uses a global var 🙀 so that you only need a single function, `show!`, without any extra setup.
    
# 2. Rename `view` fns to `show` 

for consistency with the java version and with Wolfram having a Show function.

# 3.  graphics/show! supports multiple windows/visualizations

Modify graphics/show! to take an optional `window` argument, so that you can force it to create a new window, e.g. to display two graphs next to each other for comparison.

# 4. Improved handling of paths for non-standard installations