Improvement: Turn MessageTemplate into an exception
Context
Sometimes, evaluation is not possible and Wolfram returns an error message in the form of a "MessageTemplate," which can be evaluated to get the actual user-friendly error message.

Action:
Detect this and throw an exception with the evaluated message template, and details in data.
Added a short section about error handling to Clojure primer.