Fix #124 Add support for RuleDelayed in maps (#157)
* Fix #124 Add support for RuleDelayed in maps

Context:
Sometimes, Wolfram returns an error in the form of a RuleDelayed from `"MessageTemplate"` to an expression to get a human-friendly message. F.ex. evaluating `(list (w/Interpreter "City") "San Francisco, US")` while offline returns `Association[RuleDelayed["MessageTemplate", MessageName[Interpreter, "noknow"]]]`. We should do better than throw a parse exception. It has been decided to return the delayed expression as-is and leave it to the user to evaluate it (since calling eval during parsing is ugly).

Action:
Parse RuleDelayed just as Rule, only add a metadata :wolfram/delayed, for user's information.

* Improvement: Turn MessageTemplate into an exception

Context
Sometimes, evaluation is not possible and Wolfram returns an error message in the form of a "MessageTemplate," which can be evaluated to get the actual user-friendly error message.

Action:
Detect this and throw an exception with the evaluated message template, and details in data.
Added a short section about error handling to Clojure primer.

* Move eval error detection to error-detection